Move maths builtins to their own native class
