+++ c/common.h
@@ -17,7 +17,7 @@
 #undef DEBUG_TRACE_EXECUTION
 #undef DEBUG_TRACE_GC
 
-// #define DEBUG_STRESS_GC
+#define DEBUG_STRESS_GC
 
 #define UINT8_COUNT (UINT8_MAX + 1)
