+++ c/natives.c
@@ -2,7 +2,6 @@
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
-#include <math.h>
 
 #ifdef _WIN32
 #include <windows.h>
@@ -91,6 +90,9 @@ static Value typeNative(int argCount, Value *args) {
         switch (OBJ_TYPE(args[0])) {
             case OBJ_CLASS:
                 return OBJ_VAL(copyString("class", 5));
+            case OBJ_NATIVE_CLASS: {
+                return OBJ_VAL(copyString("class", 5));
+            }
             case OBJ_TRAIT:
                 return OBJ_VAL(copyString("trait", 5));
             case OBJ_INSTANCE: {
@@ -147,172 +149,6 @@ static Value lenNative(int argCount, Value *args) {
     return NIL_VAL;
 }
 
-static Value sumNative(int argCount, Value *args) {
-    double sum = 0;
-
-    if (argCount == 0) {
-        return NUMBER_VAL(0);
-    } else if (argCount == 1 && IS_LIST(args[0])) {
-        ObjList *list = AS_LIST(args[0]);
-        argCount = list->values.count;
-        args = list->values.values;
-    }
-
-    for (int i = 0; i < argCount; ++i) {
-        Value value = args[i];
-        if (!IS_NUMBER(value)) {
-            runtimeError("A non-number value passed to sum()");
-            return NIL_VAL;
-        }
-        sum = sum + AS_NUMBER(value);
-    }
-
-    return NUMBER_VAL(sum);
-}
-
-static Value minNative(int argCount, Value *args) {
-    if (argCount == 0) {
-        return NUMBER_VAL(0);
-    } else if (argCount == 1 && IS_LIST(args[0])) {
-        ObjList *list = AS_LIST(args[0]);
-        argCount = list->values.count;
-        args = list->values.values;
-    }
-
-    double minimum = AS_NUMBER(args[0]);
-
-    for (int i = 1; i < argCount; ++i) {
-        Value value = args[i];
-        if (!IS_NUMBER(value)) {
-            runtimeError("A non-number value passed to min()");
-            return NIL_VAL;
-        }
-
-        double current = AS_NUMBER(value);
-
-        if (minimum > current) {
-            minimum = current;
-        }
-    }
-
-    return NUMBER_VAL(minimum);
-}
-
-static Value maxNative(int argCount, Value *args) {
-    if (argCount == 0) {
-        return NUMBER_VAL(0);
-    } else if (argCount == 1 && IS_LIST(args[0])) {
-        ObjList *list = AS_LIST(args[0]);
-        argCount = list->values.count;
-        args = list->values.values;
-    }
-
-    double maximum = AS_NUMBER(args[0]);
-
-    for (int i = 1; i < argCount; ++i) {
-        Value value = args[i];
-        if (!IS_NUMBER(value)) {
-            runtimeError("A non-number value passed to max()");
-            return NIL_VAL;
-        }
-
-        double current = AS_NUMBER(value);
-
-        if (maximum < current) {
-            maximum = current;
-        }
-    }
-
-    return NUMBER_VAL(maximum);
-}
-
-static Value averageNative(int argCount, Value *args) {
-    double average = 0;
-
-    if (argCount == 0) {
-        return NUMBER_VAL(0);
-    } else if (argCount == 1 && IS_LIST(args[0])) {
-        ObjList *list = AS_LIST(args[0]);
-        argCount = list->values.count;
-        args = list->values.values;
-    }
-
-    for (int i = 0; i < argCount; ++i) {
-        Value value = args[i];
-        if (!IS_NUMBER(value)) {
-            runtimeError("A non-number value passed to average()");
-            return NIL_VAL;
-        }
-        average = average + AS_NUMBER(value);
-    }
-
-    return NUMBER_VAL(average / argCount);
-}
-
-static Value floorNative(int argCount, Value *args) {
-    if (argCount != 1) {
-        runtimeError("floor() takes 1 argument (%d given).", argCount);
-        return NIL_VAL;
-    }
-
-    if (!IS_NUMBER(args[0])) {
-        runtimeError("A non-number value passed to floor()");
-        return NIL_VAL;
-    }
-
-
-    return NUMBER_VAL(floor(AS_NUMBER(args[0])));
-}
-
-static Value roundNative(int argCount, Value *args) {
-    if (argCount != 1) {
-        runtimeError("round() takes 1 argument (%d given).", argCount);
-        return NIL_VAL;
-    }
-
-    if (!IS_NUMBER(args[0])) {
-        runtimeError("A non-number value passed to round()");
-        return NIL_VAL;
-    }
-
-
-    return NUMBER_VAL(round(AS_NUMBER(args[0])));
-}
-
-static Value ceilNative(int argCount, Value *args) {
-    if (argCount != 1) {
-        runtimeError("ceil() takes 1 argument (%d given).", argCount);
-        return NIL_VAL;
-    }
-
-    if (!IS_NUMBER(args[0])) {
-        runtimeError("A non-number value passed to ceil()");
-        return NIL_VAL;
-    }
-
-
-    return NUMBER_VAL(ceil(AS_NUMBER(args[0])));
-}
-
-static Value absNative(int argCount, Value *args) {
-    if (argCount != 1) {
-        runtimeError("abs() takes 1 argument (%d given).", argCount);
-        return NIL_VAL;
-    }
-
-    if (!IS_NUMBER(args[0])) {
-        runtimeError("A non-number value passed to abs()");
-        return NIL_VAL;
-    }
-
-    double absValue = AS_NUMBER(args[0]);
-
-    if (absValue < 0)
-        return NUMBER_VAL(absValue * -1);
-    return NUMBER_VAL(absValue);
-}
-
-
 static Value boolNative(int argCount, Value *args) {
     if (argCount != 1) {
         runtimeError("bool() takes 1 argument (%d given).", argCount);
@@ -321,7 +157,6 @@ static Value boolNative(int argCount, Value *args) {
 
     return BOOL_VAL(!isFalsey(args[0]));
 }
-
 
 static Value inputNative(int argCount, Value *args) {
     if (argCount > 1) {
@@ -374,8 +209,6 @@ static Value inputNative(int argCount, Value *args) {
 }
 
 // Natives no return
-
-
 
 static bool sleepNative(int argCount, Value *args) {
     if (argCount != 1) {
@@ -440,14 +273,6 @@ static bool collectNative(int argCount, Value *args) {
 void defineAllNatives() {
     char *nativeNames[] = {
             "clock",
-            "sum",
-            "min",
-            "max",
-            "average",
-            "floor",
-            "round",
-            "ceil",
-            "abs",
             "time",
             "len",
             "bool",
@@ -460,14 +285,6 @@ void defineAllNatives() {
 
     NativeFn nativeFunctions[] = {
             clockNative,
-            sumNative,
-            minNative,
-            maxNative,
-            averageNative,
-            floorNative,
-            roundNative,
-            ceilNative,
-            absNative,
             timeNative,
             lenNative,
             boolNative,
