+++ c/object.h
@@ -13,6 +13,7 @@
 
 #define IS_BOUND_METHOD(value)  isObjType(value, OBJ_BOUND_METHOD)
 #define IS_CLASS(value)         isObjType(value, OBJ_CLASS)
+#define IS_NATIVE_CLASS(value)  isObjType(value, OBJ_NATIVE_CLASS)
 #define IS_TRAIT(value)         isObjType(value, OBJ_TRAIT)
 #define IS_CLOSURE(value)       isObjType(value, OBJ_CLOSURE)
 #define IS_FUNCTION(value)      isObjType(value, OBJ_FUNCTION)
@@ -27,6 +28,7 @@
 
 #define AS_BOUND_METHOD(value)  ((ObjBoundMethod*)AS_OBJ(value))
 #define AS_CLASS(value)         ((ObjClass*)AS_OBJ(value))
+#define AS_CLASS_NATIVE(value)  ((ObjClassNative*)AS_OBJ(value))
 #define AS_TRAIT(value)         ((ObjTrait*)AS_OBJ(value))
 #define AS_CLOSURE(value)       ((ObjClosure*)AS_OBJ(value))
 #define AS_FUNCTION(value)      ((ObjFunction*)AS_OBJ(value))
@@ -43,6 +45,7 @@
 typedef enum {
     OBJ_BOUND_METHOD,
     OBJ_CLASS,
+    OBJ_NATIVE_CLASS,
     OBJ_TRAIT,
     OBJ_CLOSURE,
     OBJ_FUNCTION,
@@ -164,6 +167,12 @@ typedef struct sObjClass {
     Table methods;
 } ObjClass;
 
+typedef struct sObjClassNative {
+    Obj obj;
+    ObjString *name;
+    Table methods;
+} ObjClassNative;
+
 typedef struct sObjTrait {
     Obj obj;
     ObjString *name;
@@ -185,6 +194,8 @@ typedef struct {
 ObjBoundMethod *newBoundMethod(Value receiver, ObjClosure *method);
 
 ObjClass *newClass(ObjString *name, ObjClass *superclass);
+
+ObjClassNative *newClassNative(ObjString *name);
 
 ObjTrait *newTrait(ObjString *name);
