+++ c/optionals/math.h
@@ -1,0 +1,11 @@
+#ifndef dictu_math_h
+#define dictu_math_h
+
+#include <string.h>
+
+#include "optionals.h"
+#include "../vm.h"
+
+void createMathsClass();
+
+#endif //dictu_math_h
