+++ c/optionals/optionals.h
@@ -1,0 +1,9 @@
+#ifndef dictu_optionals_h
+#define dictu_optionals_h
+
+#include "math.h"
+#include "../vm.h"
+
+void defineNativeMethod(ObjClassNative *klass, const char *name, NativeFn function);
+
+#endif //dictu_optionals_h
